+++ app/graphql/crud/itemsubcategories.py
@@ -1,6 +1,8 @@
 # app/graphql/crud/itemsubcategories.py
 from sqlalchemy.orm import Session
 from dataclasses import asdict
+from sqlalchemy import exists
+from app.models.items import Items
 
 from app.models.itemsubcategories import ItemSubcategories
 from app.models.itemcategories import ItemCategories
@@ -67,6 +69,13 @@ def update_itemsubcategories(db: Session, subcategoryid: int, data: ItemSubcateg
 def delete_itemsubcategories(db: Session, subcategoryid: int):
     obj = get_itemsubcategories_by_id(db, subcategoryid)
     if obj:
+        linked_items = db.query(
+            exists().where(Items.ItemSubcategoryID == subcategoryid)
+        ).scalar()
+        if linked_items:
+            raise ValueError(
+                "Cannot delete item subcategory because it is referenced by other records"
+            )
         db.delete(obj)
         db.commit()
     return obj
